@@ -5,16 +5,22 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "start": "next start"
+    "start": "next start",
+    "lint": "eslint --ext .ts,.tsx . && stylelint '**/*.scss'"
   },
   "dependencies": {
     "@typeform/embed-react": "^1.0.5",
+    "@typescript-eslint/eslint-plugin": "^5.48.1",
+    "@typescript-eslint/parser": "^5.48.1",
     "airtable": "^0.11.6",
     "autoprefixer": "^10.4.13",
     "bcrypt": "^5.0.1",
     "classnames": "^2.3.1",
     "dotenv": "^16.0.3",
-    "eslint": "^7.32.0 || ^8.2.0",
+    "eslint": "^8.31.0",
+    "eslint-config-airbnb-typescript": "^17.0.0",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.0",
     "file-loader": "^6.2.0",
     "firebase": "^8.7.1",
     "jsonwebtoken": "^8.5.1",
@@ -25,12 +31,19 @@
     "react-dom": "17.0.2",
     "react-google-login": "^5.2.2",
     "sass": "^1.34.1",
+    "stylelint": "^14.16.1",
+    "stylelint-config-recommended-scss": "^8.0.0",
+    "stylelint-scss": "^4.3.0",
     "tailwindcss": "^3.2.4"
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.19.1",
     "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.25.2",
+    "eslint-plugin-import": "^2.27.4",
     "typescript": "^4.9.4"
-  }
+  },
+  "main": "index.js",
+  "repository": "https://github.com/Coolspots/coolspots.git",
+  "author": "Pablo Grau <paugf89@gmail.com>",
+  "license": "MIT"
 }
