@@ -4,57 +4,29 @@ module.exports = {
     es2021: true,
     node: true,
   },
-  extends: 'airbnb-base',
+  parser: '@typescript-eslint/parser',
+  plugins: ['react', '@typescript-eslint', 'import'],
+  extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended', 'airbnb-typescript'],
   overrides: [],
-  parser: 'babel-eslint',
   parserOptions: {
+    tsconfigRootDir: __dirname,
     ecmaVersion: 'latest',
     sourceType: 'module',
+    project: './tsconfig.json',
   },
   rules: {
-    'react/sort-comp': 'off',
-    'react/jsx-curly-newline': 'off',
-    'react/function-component-definition': 'off',
-    'react/jsx-no-useless-fragment': 'off',
-    'no-unused-expressions': 'off',
-    'no-restricted-exports': 'off', // until restrictDefaultExports is available (https://github.com/eslint/eslint/issues/15617)
-    'import/no-dynamic-require': 'off',
-    'global-require': 'off',
-
-    'default-param-last': 'off',
-    'prefer-regex-literals': 'off',
-    'no-import-assign': 'off',
-    'no-unsafe-optional-chaining': 'off',
-    'no-promise-executor-return': 'off',
-    'react/jsx-no-constructed-context-values': 'off',
-    'react/no-arrow-function-lifecycle': 'off',
-    'react/no-unused-class-component-methods': 'off',
-    'react/no-unstable-nested-components': 'off',
-    'import/no-unresolved': 'off',
-    'import/no-relative-packages': 'off',
-    'import/no-import-module-exports': 'off',
-    'react/boolean-prop-naming': [
+    'react/jsx-filename-extension': ['error', { extensions: ['.js', '.jsx', '.tsx'] }],
+    'import/extensions': [
       'error',
-      { rule: '^(is|are|has|can|with|in|no)[A-Z]([A-Za-z0-9]?)+' },
+      'always',
+      { js: 'never', jsx: 'never', ts: 'never', tsx: 'never' },
     ],
-    'react/jsx-props-no-spreading': [
-      'error',
-      {
-        html: 'enforce',
-        custom: 'ignore',
-        explicitSpread: 'ignore',
-        exceptions: [],
-      },
-    ],
-    'no-underscore-dangle': [
-      'error',
-      {
-        allow: [
-          '__PRELOADED_STATE__',
-          '__REDUX_DEVTOOLS_EXTENSION_COMPOSE__',
-          '__APOLLO_STATE__',
-        ],
-      },
-    ],
+    semi: ['error', 'always'],
+    quotes: ['error', 'single'],
+    'no-console': ['error'],
+    'no-unused-vars': ['error'],
+    indent: ['error', 2],
+    'no-trailing-spaces': ['error'],
+    'eol-last': ['error', 'always'],
   },
 };
