@@ -1,5 +1,6 @@
-import "../styles/globals.scss";
-import { AuthProvider } from "../contexts/AuthContext";
+import '../styles/globals.scss';
+import { AuthProvider } from '../contexts/AuthContext';
+
 const App = ({ Component, pageProps }) => {
   return (
     <AuthProvider>
